Move cffi verify to cffi.set_source builders.
