+++ cffi_builders/lib/alphablit.c
