+++ pygame/_jpg.py
@@ -1,0 +1,2 @@
+from _jpg_c import ffi
+from _jpg_c import lib as jpglib
