+++ pygame/_png.py
@@ -1,0 +1,2 @@
+from _png_c import ffi
+from _png_c import lib as pnglib
